Filter requirements.
No need to say it depends on a lot of interanlly-defined shit. So override
`rpmbuild`'s internal auto requires filter, tell it to use the external
filter, and then override the external filter with our own. Our own just runs
the original external find-requires script, then filters-out its output with a
regular expression.

Also tweak some of the requires and provides, and say that we provide "iov-perl",
so that one can just depend on that without needing the Perl version prefix.

Also, add a README.