Improve error handling.
Add --die to force it to die on error, instead of continuing. Also, keep
modules and distributions separate when looking for what we've already seen.
Otherwise single-line modules like Inline, which have the same name as the
distribution, will never be installed!

Encapsulate the --die handling into a new method, vent(), which warns or dies,
as appropriate.