Separate vendor man pages from core man pages.
So that when we install a dual-life module RPM, such as Module-Build or CPAN,
there are no file conflicts. This allows us to also remoev the `--replacepkgs`
and `--replacefiles` options to `rpm` when installing just-created RPMs.

Or, well, almost. There is also the issue of dual-life bin files, such as
`config_data` (Module-Build) and `json_pp` (JSON-PP). So also update the Perl
spec file to install such dual-life bin scripts with the perl version (520)
prefixed to them, and then use `update-alternatives` to create or delete them
in %post and %postrun phases. This is as documented in the
[PackagingDrafts/UsingAlternatives](http://fedoraproject.org/wiki/PackagingDrafts/UsingAlternatives)
wiki page. It seems to work really well. After installing the perl520 RPM, you
can use `config_data`, and then you can install perl520-Module-Build and still
use it. Uninstall perl520-Module-Build and you can *still* use it. It's like
magic!

So this should eliminate RPM conflicts for dual-life modules. The only upshot
is that the list of bin files included in the Perl 5 core should be reviewed
for dual-lifeness for each major release of Perl to ensure that it is
up-to-date.