Remove Number-Phone override stuff.
The MetaCPAN API has fixed [its bug](https://github.com/CPAN-API/cpan-api/issues/364)
and reindexed Number-Phone, so it now lists the correct version. So just kill
off the whole "override" thing, as it was a nasty hack, and should only ever
be needed in cases of (I'm hoping) rare MetaCPAN bugs like this. Which is to
say: hopefully never again.