Add support for system Perl and improve diagnostics.
If no `--version` is passed, assume the current Perl is the version to build
against, and detect if it's the system Perl by checking `$Config{prefix}`. For
the system Perl, we make the following modifications:

* Make sure that the `perl-core` RPM is installed, so we get all core
  dependencies.
* Set Epoch to the minor Perl version, e.g., "16" for Perl 5.16.x.
* Set the default prefix to /usr.
* Specify an empty `plv` macro.
* Skip building mod_perl. It's a PITA and we can rely on the EPEL RPM.

In addition, module RPMs now require the minor Perl version, not the patch
level. That is, they now require "5.16" instead of "5.16.3". This should allow
them to work with any installed minor version of Perl instead of an exact
version.

The files to include an RPM are now derived from the `.packlist` file created
by the Perl module build process, rather than just a wildcard of all files and
directories. This is because directories created when building against the
system Perl conflict with files in the `filesystem` RPM. So it makes sense just
to get an exact list from the Perl build process to avoid such conflicts.

To properlly support the Epoch variable, add the `--epoch` flag. This allows
it to be set for any Perl build.

To make diagnosing failures easier, mention the name of each distribution that
fails to build in the error message. Also, collect the names of all such
failed distributions and list them all again just before the program exits, so
that you no longer have to go over all the output to find the failures.

Update `etc/dists.json` to correct for some circular dependencies and to add
some missing dependencies. Also let `DBD-Pg` depend on `postgresql` and
`postgresql-devel`, instead of a version-specific RPM.