Don't skip prereqs for different phases.
Noticed that Sereal was requiring Sereal::Encoder and Sereal::Decoder during
the test phase but not the runtime phase, even though the MetaCPAN API was
properly declaring it for both. Realized that the code was skipping a prereq
if it had already been seen, even if from a different phase. So now add a
previously seen prereq if it's from a different phase, but build it only the
first time it's seen.