Add version_format configuration field.
This allows one to specify a format for RPM versions, as well as the main
module provided by the RPM. Necessary because RPM employs a dot-segmented
`cmp` when comparing versions, and sometimes a CPAN module might be released
as 3.0014 and succeeded by 3.1. In that case, RPM thinks that 3.0014 is newer,
since it reads it as, essentially, 3.14. So add `version_format`, with a
format like "x.xxxx". This causes the RPM representation of version 3.1 to be
formatted as 3.1000". The format also affects the formatting of the main
module version in the "provides" metadata of the RPM, but no other module
versions are affected.

While this can cause confusion for someone who requires
`perl(Number::Phone) = 3.1` and is suprised when 3.0014 satisfies that
requirment, this change is required to prevent RPM from incorrectly
complaining about a newer version when upgrading. Overall, it's intended to be
a workaround for the relatively rare case when a module is released with fewer
digits after a decimal point. In such a case, it could pay to file a bug
report asking for the 0-padding to be restored in the next release, since it
will continue to be valid for numeric comparisons, as well.

Number-Phone exhibited this challenge, hence this patch and the use of
`version_format` in the configuration for that distribution.

Also, make sure to require a modern version of version.pm. Probalby not
strictly necessary, but prudent.