Report distribution a module is found in.
Also, indent the information for a module a couple of spaces. These changes
makes it easier to see what module caused a distribution to be built.
Important for when MetaCPAN misreports the distribution a module is in, which
has happend a few times, unfortunately
([example](https://github.com/metacpan/metacpan-api/issues/614)).