Add `dirname` field.
As well as the patch and version format necessary to fix the IO-Pager 0.40
release. Because the version is declared as a number, the trailing "0" gets
removed from the directory name and the version in the META files. Just
setting the name to "IO-Pager-0.4" doesn't work, because then we're not able
to query the MetaCPAN API for the modules provided by this distribution. So
add the `dirname` field so that we can properly `cd` into the proper directory
but don't change any other parts of the name. Then set the version format and
patch for IO-Pager 0.40 so that it builds properly with stringified versions.